backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-sessions==0.3.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
loguru==0.7.3
mysql-connector-python==9.1.0
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
win32_setctime==1.2.0
bcrypt
passlib[bcrypt]
